+++ sensemap-backend/seeds/dev/00_clear.ts
@@ -1,16 +1,15 @@
 import * as Knex from "knex";
 
-exports.seed = function(knex: Knex): Promise<any> {
-  return knex("edge")
-    .del()
-    //.then(() => knex("transaction").del())
-    .then(() => knex("history").del())
-    .then(() => knex("annotation").del())
-    .then(() => knex("object").del())
-    .then(() => knex("card").del())
-    .then(() => knex("box").del())
-    .then(() => knex("map").del())
-    .then(() => knex("oauth_authorization_code").del())
-    .then(() => knex("oauth_token").del())
-    .then(() => knex("user").del());
+exports.seed = async function(knex: Knex): Promise<any> {
+  await knex("edge").del();
+  await knex("transaction").del();
+  await knex("history").del();
+  await knex("annotation").del();
+  await knex("object").del();
+  await knex("card").del();
+  await knex("box").del();
+  await knex("map").del();
+  await knex("oauth_authorization_code").del();
+  await knex("oauth_token").del();
+  await knex("user").del();
 };