+++ sensemap-backend/src/types/history.ts
@@ -1,5 +1,5 @@
 import * as Knex from "knex";
-import { Transaction } from "./transactions";
+import { Transaction } from "./transaction";
 
 export async function writeHistory(db: Knex, trx: Transaction) {
 }